requirements.txt
pytesseract
Pillow
bottle
